Revert "[Tests] Convert reduce tests to gTests (#2848)"
This reverts commit 88720ba1800331a30d0478d90d8f0c25330d390b.
